+++ src/main/java/mg/itu/aquanova/alimentation/models/PrevisionResult.java
@@ -1,22 +1,21 @@
 package mg.itu.aquanova.alimentation.models;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class PrevisionResult {
 
     private Long alimentId;
     private String alimentNom;
-    private BigDecimal stockRestant;
-    private BigDecimal consommationParJour;
+    private Double stockRestant;
+    private Double consommationParJour;
     private Long joursRestants;
     private LocalDate dateRupture;
-    private BigDecimal quantiteSuggereePurchase;
+    private Double quantiteSuggereePurchase;
     private String alerte;
 
     public PrevisionResult() {}
 
-    public PrevisionResult(Long alimentId, String alimentNom, BigDecimal stockRestant,BigDecimal consommationParJour, Long joursRestants,LocalDate dateRupture, BigDecimal quantiteSuggereePurchase) {
+    public PrevisionResult(Long alimentId, String alimentNom, Double stockRestant, Double consommationParJour, Long joursRestants, LocalDate dateRupture, Double quantiteSuggereePurchase) {
         this.alimentId = alimentId;
         this.alimentNom = alimentNom;
         this.stockRestant = stockRestant;
@@ -32,11 +31,11 @@ public class PrevisionResult {
     public String getAlimentNom() { return alimentNom; }
     public void setAlimentNom(String alimentNom) { this.alimentNom = alimentNom; }
 
-    public BigDecimal getStockRestant() { return stockRestant; }
-    public void setStockRestant(BigDecimal stockRestant) { this.stockRestant = stockRestant; }
+    public Double getStockRestant() { return stockRestant; }
+    public void setStockRestant(Double stockRestant) { this.stockRestant = stockRestant; }
 
-    public BigDecimal getConsommationParJour() { return consommationParJour; }
-    public void setConsommationParJour(BigDecimal consommationParJour) { this.consommationParJour = consommationParJour; }
+    public Double getConsommationParJour() { return consommationParJour; }
+    public void setConsommationParJour(Double consommationParJour) { this.consommationParJour = consommationParJour; }
 
     public Long getJoursRestants() { return joursRestants; }
     public void setJoursRestants(Long joursRestants) { this.joursRestants = joursRestants; }
@@ -44,17 +43,17 @@ public class PrevisionResult {
     public LocalDate getDateRupture() { return dateRupture; }
     public void setDateRupture(LocalDate dateRupture) { this.dateRupture = dateRupture; }
 
-    public BigDecimal getQuantiteSuggereePurchase() { return quantiteSuggereePurchase; }
-    public void setQuantiteSuggereePurchase(BigDecimal quantiteSuggereePurchase) { this.quantiteSuggereePurchase = quantiteSuggereePurchase; }
+    public Double getQuantiteSuggereePurchase() { return quantiteSuggereePurchase; }
+    public void setQuantiteSuggereePurchase(Double quantiteSuggereePurchase) { this.quantiteSuggereePurchase = quantiteSuggereePurchase; }
 
-    public BigDecimal getQuantiteRestante() { return stockRestant; }
-    public void setQuantiteRestante(BigDecimal quantiteRestante) { this.stockRestant = quantiteRestante; }
+    public Double getQuantiteRestante() { return stockRestant; }
+    public void setQuantiteRestante(Double quantiteRestante) { this.stockRestant = quantiteRestante; }
 
-    public BigDecimal getConsommationJour() { return consommationParJour; }
-    public void setConsommationJour(BigDecimal consommationJour) { this.consommationParJour = consommationJour; }
+    public Double getConsommationJour() { return consommationParJour; }
+    public void setConsommationJour(Double consommationJour) { this.consommationParJour = consommationJour; }
 
-    public BigDecimal getSuggestionQuantiteAchat() { return quantiteSuggereePurchase; }
-    public void setSuggestionQuantiteAchat(BigDecimal suggestionQuantiteAchat) { this.quantiteSuggereePurchase = suggestionQuantiteAchat; }
+    public Double getSuggestionQuantiteAchat() { return quantiteSuggereePurchase; }
+    public void setSuggestionQuantiteAchat(Double suggestionQuantiteAchat) { this.quantiteSuggereePurchase = suggestionQuantiteAchat; }
 
     public String getAlerte() { return alerte; }
     public void setAlerte(String alerte) { this.alerte = alerte; }
