+++ src/main/java/mg/itu/aquanova/referentiel/models/EspecesModels.java
@@ -1,7 +1,5 @@
 package mg.itu.aquanova.referentiel.models;
 
-import java.math.BigDecimal;
-
 import jakarta.persistence.Table;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -20,7 +18,7 @@ public class EspecesModels {
     private String nom;
 
     @Column(name = "poids_cible_moyen", precision = 6, scale = 2)
-    private BigDecimal poidsCibleMoyen;
+    private Double poidsCibleMoyen;
 
     @Column(name = "image", columnDefinition = "TEXT")
     private String image;
@@ -44,11 +42,11 @@ public class EspecesModels {
         this.nom = nom;
     }
 
-    public BigDecimal getPoidsCibleMoyen() {
+    public Double getPoidsCibleMoyen() {
         return poidsCibleMoyen;
     }
 
-    public void setPoidsCibleMoyen(BigDecimal poidsCibleMoyen) {
+    public void setPoidsCibleMoyen(Double poidsCibleMoyen) {
         this.poidsCibleMoyen = poidsCibleMoyen;
     }
